O3-1608: Set "Reason for Change" when substituting drugs
